program counter u32 -> u16

+++ src/class/code.rs
@@ -16,8 +16,8 @@ impl Instruction {
         Self::new(Opcode::OperationSpacer, vec![])
     }
 
-    pub fn size(&self) -> u32 {
-        1 + self.operands.len() as u32
+    pub fn size(&self) -> u16 {
+        1 + self.operands.len() as u16
     }
 }
